Added voting.timer.ts
